Pass install user into bootstrap and runserver scripts.
Remove hard-coding of "vagrant" user and, instead, pass it in to the
bootstrap.sh and runserver scripts.

The bootstrap.sh script now expects these arguments:
- User under which to install application.  This user must already
exist and is expected to have a home directory under /home;
- Name of Hydra head;
- Name of GitHub repo containing Hydra head application

The bootstrap.sh script contains defaults suitable for AWS.  The
Vagrantfile overrides these with values suitable for Vagrant.

The runserver script now takes the install user as its first argument
and the name of the Hydra head as its second.