Run bootstrap script as root.
The bootstrap.sh script is now run as a privileged user, i.e., root.
This is to make it easier to use the same script for bootstrapping AWS
instances, which run bootstrapping scripts as root.

Commands that previously required "sudo" to run them as root have had
the "sudo" removed.  Likewise, commands that previously ran as the
unprivileged user are run via "sudo," specifying the unprivileged user
as the one to run under.

The "-H' flag is used in the sudo command to ensure $HOME is set
correctly.  This prevents problems with Bundler, which stores things
under $HOME/.bundler.  Without the "-H", Bundler would fail trying to
create files under root's home directory.