Replace bundled hydra-jetty with standalone Solr and Fedora 4 services.
This change alters the installation script so that hydra-jetty is no
longer installed and used.  Instead, a separate Solr and Fedora 4 are
installed and used by the application.

Currently, we install Solr 5.2.1 and Fedora 4.2.0.  Fedora is built
from a specific source branch, as per request.  The Solr installation
uses the install_solr_service.sh script that comes bundled with Solr 5.
 Fedora is installed to be run via Tomcat 7.

Tomcat 7 runs on port 8080 and Solr on port 8983.  Solr installs under
/opt/solr and runs under user "solr".  Its cores and related server
runtime data are kept under /var/solr.  Fedora's data is stored under
$installuser/fedora-data.

Here are some specific changes:

 - Hydra-jetty is no longer started explicitly in the Vagrantfile
runserver script.

 - Local port 8888 is forwarded to the Tomcat 7 service (port 8080) in
the Vagrantfile.

 - The major version number in bootstrap.sh has been bumped.

 - The version of Rails that is installed has been locked to Rails
4.2.x.

 - Solr installs under its own user, as per the install_solr_service.sh
script.  It was kept at the default "solr" so as better to isolate Solr
from the application.

 - Beware: Fedora 4 installs with the default fedoraAdmin password.

 - Fedora 4 is deployed to Tomcat 7 as a WAR file built from source.

 - Hydra-jetty is no longer installed, configured, and started.

 - The path to the FITS executable is now set explicitly in
config/initializers/sufia.rb.

 - Solr cores are created for the application.  The core name
corresponds to the Rails environment under which the application is
running, i.e., development or production.

 - Application configuration files config/solr.yml,
config/blacklight.yml, and config/fedora.yml are patched to point to
the standalone Solr and Fedora services.

 - Solr and Tomcat services are stopped after install whilst the
application is being installed and configured.  Their services are
started after everything is installed.