Backport improvements from ETDplus development
Several improvements have been made during the development of the
ETDplus application installation process.  We backport the improvements
to those scripts to the InstallScripts used by data-repo.

A major change is that, when installing, we do a straightforward "git
clone" of the data-repo code instead of the current process that
involves "rails new".  We also pass in a config/secrets.yml file during
installation that either is user-supplied or auto-generated.

Although we don't currently use private repos with data-repo, the
support for SSH deploy keys has also been backported.