Remove some configuration fixups.
One of the last steps in installing the Sufia application is to
fix up Solr core names in config/solr.yml and config/blacklight.yml,
and also to amend the listening port for Fedora in config/fedora.yml.
These values have since been amended in the application source
code and so these steps are now effectively a no-op.  This commit
removes themi from the installation process.