Incorporate GeoBlacklight support into Ansible playbooks
This adds support for installing the GeoBlacklight project via the
Ansible playbooks.

GeoBlacklight support adds a number of roles, including: geoblacklight;
chroot-sftp; and tls_cert.  We also add a cron job to run the rake
task vtul:geoblacklight:data_ingest each day at midnight.

The Vagrantfile has been modified to support the selection of which
application type to install and which provisioner to be used.  The
defaults are "sufia" and "ansible_local" respectively.

The ANSIBLE_PROVISIONER environment variable may be set prior to
"vagrant up" to select which version of the Vagrant Ansible provisioner
is used: ansible or ansible_local.

If the APP_TYPE environment variable is set prior to "vagrant up" then
it is used to determine which type of application to provision if it is
set to a known type: "sufia" or "geoblacklight".  If it is not set, or
is not set to a valid application type then the value used is
determined by examining the "project_name" setting in the
ansible/site_secrets.yml file.