Refactor YAML file processing and handle exceptions
Currently, Vagrant processes the "ansible/site_secrets.yml" file as
part of parsing the Vagrantfile.  If the "ansible/site_secrets.yml"
file has not been created then an exception will be raised, even if
apparently benign commands such as "vagrant box update" or "vagrant
status" are being executed.

This commit adds exception handling to the YAML file processing so as
to issue a reminder warning if the "ansible/site_secrets.yml" file is
missing instead of exiting due to a fatal exception.

Also, this commit refactors the YAML file processing so that it is done
only once and the results stored in a global hash instead of processing
the file every time site_file() is invoked (which is several times).
Making the hash global also makes the "ansible/site_secrets.yml"
settings potentially available elsewhere in the future.