Add Ansible role to install FITS
This adds a "fits" role to install FITS for Sufia applications and adds
it as a dependency of the "sufia" role.  FITS is installed to /opt and
the version installed is symlinked to /opt/fits (as in the previous
BASH script InstallScripts).

Version 0.8.5 of FITS is installed for IAWA (i.e., Sufia 7); version
0.6.2 otherwise.  This can be overridden by setting "fits_version" to
the desired version in site_secrets.yml.