Fix data_ingest cron job
The PATH used in cron jobs is more restricted than when running the
same commands directly from a login shell.  This was causing "bundle
exec" in the cron job to fail due to "bundle" not being on the PATH.
The fix here is to invoke the bundle command with an absolute pathname,
i.e., "/usr/local/bin/bundle exec".