Support ClamAV PrivateMirrors
Add support to explicitly add a PrivateMirror directive to
/etc/clamav/freshclam.conf if the Ansible setting
clamav_private_mirror_enabled is set to true.  If set to false, the
PrivateMirror option is removed from /etc/clamav/freshclam.conf.  The
Ansible setting clamav_private_mirror_host designates the hostname of
the PrivateMirror to be used.

The default is to enable the VT ClamAV PrivateMirror.  This should
speed up the initial freshclam invocation considerably.