Add additional settings to define which ORCiD service is used
Here we add extra fields to the "orcid" key of the config/secrets.yml
template so that either the "sandbox" or "production" ORCiD services
may be used.

Which settings are used is governed by the "project_orcid_service"
entry in site_secrets.yml.  URLs corresponding to the production
service are used if "project_orcid_service" is set to "production".
Otherwise, settings appropriate for the sandbox service are used.