Change APT caching behaviour for Ansible 2.2
With Ansible 2.2, tasks that use the "apt" module now appear to update
the APT cache unconditionally when used with the "update_cache: yes"
option.  A side-effect of this is that the task is always flagged as
"changed" during an Ansible playbook run.  This "changed" status
applies even if no new packages are installed by the "apt" module.  In
a sense, such "changed" statuses can be seen as false positives in the
PLAY RECAP tally.

This pull request change is twofold:

1) It adds "update_cache: yes" options to tasks using "apt_repository",
where arguably they are more needed.  The "update_cache" option appears
to act more sensibly with "apt_repository" in that the status of
"changed" is applied and the APT update only done if the APT repository
is actually added/updated.

2) we add a "cache_valid_time" option to tasks using the "apt" module
so that an "apt update" requested by "update_cache: yes" is only
performed if the APT cache is older than "apt_cache_timeout" seconds.
The "apt_cache_timeout" value is a global set in ansible/site_vars.yml
and defaults to 30 minutes.  This means "update_cache: yes" will no
longer fire unconditionally, meaning tasks using the "apt" module will
not always register as "changed", even when no actual packages were
installed (e.g., on a re-provision).

With this change, the total number of "changed" statuses was reduced
from 17 to 5 on a re-provision without any explicit changes (restoring
the previous Ansible 2.1 behaviour).

Note: by setting "apt_cache_timeout" to a very short amount (e.g., 1
second), the effect of unconditional APT updates may be restored, if
desired.