Stop installing Gems as root
Previously, application Gems were installed system-wide, which required
elevated privileges.  This change installs Gems locally to the
vendor/bundle directory as an unprivileged user.

As part of the change, permissions were made less permissive.
Previously, the application owned all the files and directories.
Should the code be exploitable, it would be fairly straightforward to
overwrite/modify application code as part of the exploit.  Thus, this
change introduces a new user (the application user) under which the
application runs.  The previous user (now known as the application
owner) now owns the files and the application user has only limited
permissions to write to certain directories that are necessary for the
application to run (e.g., log/, tmp/, etc.).  The application user no
longer has write access to application code files, which hopefully
should improve runtime safety.