Fix Passenger setup
The current Nginx Passenger setup is designed to work with only a
single Rails application.  Because sites-enabled currently uses
{{project_name}} it is possible to get multiple Passenger sites enabled
if the project_name is changed.  This, in turn, will cause Nginx to
fail to start because of complaints about multiply-defined
Passenger-related settings.

This change uses a single Passenger site name in sites-enabled, thereby
enforcing only one to the enabled at any given time when deployed via
these Ansible scripts.