Support local or remote PostgreSQL servers for Rails DB
The original BASH scripts supported the use of a local or remote
PostgreSQL server for the Rails ActiveRecord DB.  The Ansible version
was a regression in that it supported only the use of a local
PostgreSQL server.  This commit eliminates this regression and allows
either a local or a remote PostgreSQL server to be used.

If {{project_db_host}} points either to "localhost" or a directory (in
which the PostgreSQL Unix socket resides) then it is assumed a local
PostgreSQL server is to be used.  In this case a local server is
installed and set up.  Otherwise, it is assumed a remote PostgreSQL
server (already installed) is to be used, in which case only necessary
PostgreSQL client libraries are installed.