Fix cron job timing specification (#131)
The cron job to process SFTP uploads is supposed to run once a day at
midnight. The current time specification to cron is incorrect and
causes the job to run every minute between midnight and 1 am.  This
change fixes that.

Also, we silence the output of the cron job as the rake task logs
everything to files under the SFTP "Reports" directories.  This avoids
harmless warnings being mailed by cron each day.