Make explicit the "graylog_enable" setting (#136)
The code to toggle on or off logging to Graylog is governed by the "graylog_enable" setting in the Ansible playbooks.  Although this is set by default in the appropriate roles, it has always been possible to override the default setting by setting "graylog_enable" explicitly in "site_secrets.yml".  This change documents that ability explicitly in "example_site_secrets.yml".