Support sidekiq as an Active Job backend
Add support for choosing which Active Job backend to use: resque or sidekiq. The variable "active_job_backend" designates which backend to use. Currently, it must be set either to the string "resque" or "sidekiq". It defaults to "sidekiq" currently for the "iawa" project, otherwise it defaults to "resque".

For the "sidekiq" backend, the variable "sidekiq_workers" controls the number of sidekiq processes to spawn to handle backend task processing. It defaults to 1. For large amounts of concurrent background processing, this should be increased. Each sidekiq process typically controls 25 background processing threads.