Make sure Fedora 4 and Solr are running before rake tasks run
Modified the "fedora4" role so that it starts Tomcat again at the end of its tasks having stopped it at the beginning of them. Also, modified the "solr" role to have the service in a running state at the end of its tasks (this was not explicit before). Finally, had Ansible flush all handlers if there are any project-local tasks to be run in the "sufia" role. This ensures that any Solr cores, etc., defined earlier are reflected in the running services (which are restarted by the handlers to do so). That way, any rake tasks that rely on functioning Fedora 4 and Solr services will work.

Because of the above, various rake tasks that were previously commented out (because they failed during the Ansible provisioning) have now been uncommented (at least in the case of "iawa"). This resulted in bringing some permissions problems to light due to the way the project-local tasks were run. Making the project-local tasks be more granular in which user was used for each task solved the permissions problem.