Make it easier to switch between Java roles
Currently, selecting whether to use the "java8" (Oracle) or "openjdk8" (OpenJDK) roles involves changing some code in the roles that use Java. This change makes the choice of Java role more easily selectable via the variable "java_vendor". If "java_vendor" is set to "oracle" then the "java8" role is used. If it is set to "openjdk" then the "openjdk8" role is used.

The default for "java_vendor" is currently "openjdk".