Ensure "assets:precompile" is done in the correct place
Currently, the "assets:precompile" Rake task that is done for RAILS_ENV=production mode is done right at the very end of the deployment process.  Unfortunately, any application-specific Rake tasks that utilise the asset pipeline will fail in RAILS_ENV=production mode (at the very least on a fresh install) because these assets are not in place due to the tasks being run before the "assets:precompile" task is reached.

This change makes the application-specific Rake/provisioning tasks run at the end of provisioning.