Support Hyrax via a role (#140)
* Support Hyrax via a role

Up to now, Sufia and Hyrax Samvera applications have both used the "sufia" role in order to provision. This has lead to problems due to the subtle differences between the two.

This change makes "hyrax" an explict role. We also amend our Vagrant setup so that that application type now defaults to "hyrax" and we add a new top-level site file for building Hyrax applications.