Allow Riiif cache duration to be passed via config/secrets.yml
This change allows the Riiif cache_duration parameter to be set via the application's config/secrets.yml file instead of having it be hard-coded in the config/initializers/riiif.rb initializer itself.

This adds a new hierarchy to config/secrets.yml: iiif.  The cache_duration is a key, "cache_duration", which represents to amount of time (in seconds) an object will remain in the Riiif cache.