Support local /etc/hosts overrides via Vagrant "vagrant-hostmanager" plugin
Use the Vagrant "vagrant-hostmanager" plugin to make it easier to manage local hostname overrides for Vagrant VMs.

When this plugin is installed, a local hostname of "<project_name>.dld.lib.vt.edu" is inserted into the host system's /etc/hosts file and may be used to access the application through a browser on the local machine.

These /etc/hosts overrides are useful for allowing applications that use VT CAS authentication to work.