Fix FITS version defaults (#164)
Hyrax 2.x applications are recommended to use FITS 1.0.5.  The current code chose a default based upon "project_name" that allowed only the possibility of version 0.8.5 or 0.6.2.

This change corrects the setting of the default value of FITS so it is appropriate for our current projects.  Note: "fits_version" can always be explicitly overridden in the "ansible/site_secrets.yml" file when deploying.