Sunset EZID settings
The previous DOI provider, EZID, had config/secrets.yml referenced under the key space of "ezid".  When this was changed to the more generic key space of "doi", older applications that still referenced "ezid" broke due to the absence of that key hierarchy in config/secrets.yml.

This change restores the old "ezid" settings alongside the new, preferred "doi" settings, for backwards compatibility with applications that have yet to update to using "doi".

Note that the "ezid" keys are now deprecated and should be removed in the future.