Make "clamav" role work for redeployments
The "clamav" role is designed to ensure that virus definitions are present when deploying so that Samvera applications will subsequently start up. (If virus definitions are not present, the ClamAV Gem will fail to initialise.)

To ensure virus definitions are present and usable, the "clamav" stops the "clamav-freshclam" service; invokes "freshclam" explicitly to create/update the virus definition files; and then starts the "clamav-freshclam" service again. The "freshclam" command is invoked using the Ansible "command" module.

Unfortunately, despite information in the "freshclam" man page to the contrary, "freshclam" returns a return code of 1 (not 0) if the virus definition files are updated or up to date. This return code is interpreted as a command failure by Ansible, and it considers the play task to have failed.

This error is normally only provoked when reprovisioning. On a fresh deployment, installation of ClamAV itself causes initial virus definitions to be placed in /var/lib/clamav and the "creates" argument to the "freshclam" invocation via "command" causes the "freshclam" update to be skipped. Subsequent updates via the "clamav-freshclam" service normally result in updated files with ".cld" extensions, which will cause the "freshclam" command to execute, as it checks for the presence of a "/var/lib/clamav/main.cvd" file.

This change fixes the "freshclam" command module execution by considering only return codes > 1 to be an error.  It also removes the "creates" argument that causes the update to be skipped when any virus definitions are present, as these may be very stale: it is the responsibility of "freshclam" to decide whether or not updates are needed.

Also, this update explicitly enables the "clamav-freshclam" service to run at system boot.  Previously, this was an implicit side-effect of package installation.