Fix Ansible 2.7 apt/pip deprecations (#179)
Ansible 2.7 now deprecates the use of loops to install multiple packages with "apt" or "pip".  Instead, the preferred method is to supply a list of packages when multiple packages are being installed.

This commit changes all uses of loops ("with_items" constructs) in conjunction with the "apt" or "pip" modules to use lists or list variables instead.  The deprecated loops feature will be removed entirely in Ansible 2.11.