Support locally-supplied Fedora WAR file
Up to now, only official distribution Fedora WAR files can be used when installing Fedora 4.  In some cases, it is desirable to use a locally-patched or otherwise locally-produced Fedora WAR file instead.

This change allows a Fedora WAR file to be placed in the "local_files" directory and used in preference to the distribution version when installing Fedora 4.  If a local WAR file is not present, it will fall back on using the distribution version, as normal.

The "fedora_war_filename" Ansible variable denotes the name of the Fedora WAR file that is used, if present, in "local_files/".