Tighten up ownerships and permissions
Tighten up the ownerships and permissions of the files unpacked from the archive.  In particular, we want to make sure the files are read-only.

This also makes the "unarchive" task run only in the event of a successful copy of the carousel package of files.

In addition, we supply a fall-back default for the "carousel_pkg" variable in case it is not defined in "site_secrets.yml".  This is a minor convenience to avoid the "copy" task failing due to an undefined variable rather than the copy itself failing.