Remove top-level "carousel" directory from carousel package
Previously, the user-supplied "carousel.tar.gz" file was expected to contain a top-level "carousel" directory under which the image directories resided.  This opens up the possibility of a user-supplied carousel package accidentally or maliciously overwriting content elsewhere in the "app/assets/images/vtul" directory hierarchy.

This change alters the semantics so as to remove the expectation of a top-level "carousel" directory in the package of carousel images.  Instead, it is expected to contain only the images and subdirectories related to the carousel, and these will be unarchived explicitly into the carousel directory during deployment.