Set random password for "upload" user
In some deployment environments the lack of a password being set for the "upload" user causes SFTP logins to be denied due to the account being considered locked.

This change sets a random password for the "upload" user.  This "unlocks" the account by setting a password but does not feasibly allow the password to be used for logins (or any other purpose that requires password authentication).  This does not affect pubkey authentication for SFTP.