Default to Ubuntu 18.04 for Vagrant
InstallScripts supports deploying to three versions of Ubuntu: 14.04, 16.04, and 18.04.  When deploying via Vagrant, the version to use can be set explicitly by setting the VAGRANT_VM_BOX environment variable to a supported Vagrant box prior to executing "vagrant up".  If VAGRANT_VM_BOX is not set then a default is used.

This change switches the default to an Ubuntu 18.04 Vagrant box from the current Ubuntu 16.04 box that is used.