Wire down uid/gid for users
InstallScripts adds several users to the system during deployment: two for Rails applications and one as an SFTP uploader (for "geoblacklight").  Currently, only the user and group names are specified when creating these users.  The actual uid and gid used is auto-assigned by the system so as not to conflict with existing users.  If users have been created on the OS outside InstallScripts then this makes the actual uid and gid settings of the InstallScripts users unpredictable.  This, in turn makes it difficult to align uid and gid assignment for, e.g., NFS mounts or other situations where uid and gid settings must match precisely.

This change allows uid and gid to be specified precisely for InstallScripts-added users.  (Note, this does not apply to users created by installed packages, e.g., Solr, Tomcat, etc., which manage their own associated users.)  The defaults for uid and gid start at 60000, which lies outside the range of system auto-assigned uid and gid, so as not to conflict.

The settings in "example_site_secrets.yml" have been updated to values that will work correctly for all currently-supported InstallScripts applications.