Add support for smart relay host (#193)
Amend the "postfix" role to allow a smart relay host to be defined. The
variable "mail_relayhost", if defined, is used to define the Postfix
"relayhost" parameter. If "mail_relayhost" is not defined then any
definition of "relayhost" is removed from the Postfix configuration
(and Postfix sends e-mail directly instead of via a smart host).

Signed-off-by: Paul Mather <paul.mather@vt.edu>