Ensure local files are unpacked for "fedora4" role
Now that the "fedora4" role has the possibility of using a locally-supplied
Fedora WAR file we need to ensure that these local files are present
before the "fedora4" role tasks are executed. This is especially true in
the case where local files are supplied via a "local_files_archive" archive.

This change corrects this omission by making the "local_files" role
a dependency of the "fedora4".

Signed-off-by: Paul Mather <paul.mather@vt.edu>