Improve switching of Ruby version (#195)
The repository from which we install Ruby supports multiple versions. We
allow the version to be selected by setting the "ruby_version" variable.
Unfortunately, changing the value of "ruby_version" does not mean the
new version of Ruby is used. This change fixes that behaviour, ensuring
explicitly that the version of Ruby selected will be used by the relevant
running Ruby applications.

The "ruby-switch" utility is used with the packaged versions of Ruby to
determine which version of Ruby installed will be the default system
Ruby. Previously, whichever version of Ruby was installed first would
become the version of Ruby used from there on. Using "ruby-switch"
after installing (or changing) Ruby versions ensures the latest
installed version will be used as default.

In addition, when the version of Ruby is switched, we need explicitly to
restart any running Ruby-based daemons/services. This change introduces
an explicit test of whether the current version of Ruby being deployed
changes from that which was already deployed. If this is the case, a new
fact is set that may be tested in other Ruby-consuming roles to know
whether or not to restart long-running daemons or services that use Ruby
installed by these playbooks.

Signed-off-by: Paul Mather <paul.mather@vt.edu>