Ensure latest Brightbox PPA Ruby version is used
The Brightbox PPA provides many different packaged versions of Ruby.
Ubuntu 18.04 (Bionic) comes with its own version of Ruby 2.5. This
creates a problem when attempting to use Ruby 2.5 from the Brightbox
PPA when using the "ruby" role in InstallScripts (e.g., by setting
"ruby_version" to "2.5" in "site_secrets.yml"). Because there is already
a "ruby2.5" package installed, the "install Brightbox Ruby package" play
invoking the "apt" module does not install "ruby2.5" from the Brightbox
PPA just added. This causes the subsequent "ruby-switch" command
to set the default to fail with the complaint that there are no
alternatives available.

This change instructs the "apt" play to install the latest versions of
the Ruby packages. This has the effect of installing all of the versions
from the Brightbox PPA.