Various deprecation removals and improvements
These changes are a potpourri of fixes to address deprecation warnings
and introduce quality-of-life improvements for developers. Changes
include the following:

- Ansible 2.8 issues a deprecation warning regarding using a "bare"
variable in a True/False context (e.g., "when: somevar"). This quietens
such deprecations by explicitly casting such variables using the "|bool"
filter.

- Changes the "hostmanager.include_offline" setting for the Vagrant
hostmanager plugin to keep track of only active VMs. This is to attempt
to remedy /etc/hosts from growing out of control.

- The "local caching" feature of the "fits" and "phantomjs" roles would
unconditionally copy ("cache") the FITS/PhantomJS distribution back to
the local Ansible control host (i.e., into "ansible/local_files")
whether or not anything had been downloaded remotely and should be
cached locally for the future. This change only copies back the FITS/
PhantomJS distribution downloaded if a local cached copy was not used.
This eliminates the time spent copying the relevant distribution back
from the system/VM being provisioned when it is unnecessary.

Signed-off-by: Paul Mather <paul.mather@vt.edu>